Fix #1617, #2745: Remove source code genrules (#5702)
## Explanation

Fixes #1617
Fixes #2745

This PR removes all ``genrule``s
(https://bazel.build/reference/be/general#genrule) that generate Kotlin
or Java source code in the repository. This corresponds to almost every
app/ directory test and production file, including the few Java files
used for binding adapters. While not every file was being regenerated,
the vast majority were.

This change has fairly substantial implications:
- Symbolization in Android Studio with the IntelliJ Bazel plugin should
now work correctly if the corresponding targets are included in one's
``.bazelproject`` file. This didn't work before because the ``genrule``s
would create a new file either with a different name or different path,
so the plugin didn't know to associate the two during sync.
- All old crash stack traces in crash dashboards (i.e. Crashlytics and
the Play Console) _may_ no longer match to the same crashes in versions
of the app that will include this change if those stack traces include
any affected app classes (since the names are now different). This isn't
certain however, since the class names are the same (only the filenames
are different, and it's possible the ``SourceFile`` attribute will be
stripped or not included in corresponding Dex files that are deployed
with the app).
- The build will be a bit faster and take up fewer resources. While the
``genrule``s themselves ran very quickly, the bigger win is the large
decrease in targets that need to be analyzed during a large build.
- Imports will now appear fairly inconsistent across contexts as we now
sometimes import ``R`` locally, from a common package, or from a very
specific package (such as for tests). Databinding imports will face a
similar issue. With symbolization in Android Studio, I'm hopeful this
isn't too confusing to developers. In the long-term, we will try to
remove the need for explicit imports entirely (and do away with
databinding), but this sort of work will be on the order of years, not
weeks or months, to complete. This will be the medium- to long-term
state for development moving forward.

This change also includes a few noteworthy simplifications:
- ``app_test`` has been removed entirely in favor of using
``oppia_android_test`` directly. This is the direction we want to go in
general for simpler test definitions, so it's nice to see this was
actually possible with the removal of ``test_with_resources``. Using
this for resource cases _is_ a bit awkward right now due to needing to
include the ``DataBinderMapperImpl.java`` bridge class for databinding,
but hopefully this will be fixable in the future (#1683).
- A bunch of structures were simplified in ``app/BUILD.bazel`` due to
the extra resources-specific splits of different libraries being no
longer necessary. The notion of migrated targets were also removed since
they were actually never needed with how Bazel's ``glob`` behaves.
- Non-modularized app tests were able to be consolidated to one
statement for both ``test/`` and ``sharedTest/`` tests (rather than
having two), which is cleaner.

There was one wiki change needed since ``app_test`` is no longer
available to use, and it has been clarified to better explain how to set
up a test for resources vs. not.

For compatibility, this further cements Gradle no longer being able to
be used for builds, but that's no longer a concern since we can't build
with Gradle, anyway.

It's important to note that all of the updated import statements are
reflecting the existing reality of the Bazel build configuration (i.e.
these are exactly the import statements that Bazel was using via the
``genrule`` targets). Android resource processing is very complex, and
even more so in the Bazel world. This means a lot of the ``R`` symbolic
references are actually duplicated many times in the build pipeline, so
there are other ``R`` files that can be used. This was an approach I
took a few months ago to try and remove the ``genrule``s while
maintaining Gradle compatibility. I didn't take that approach here for
simplicity (it was more important to get this change in), but it's quite
likely the import statements will see more iterations as we continue
with modularization work. One useful goal may be to try and consolidate
to local package ``R`` files to avoid needing an import at all (and thus
remove potential confusion that there _are_ many ``R`` files generated
throughout the build graph), but there are other possibilities that can
be explored as well.

#5747 is also tracking some future clean-up work made possible after
this change in order to simplify `R` imports throughout the codebase.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- It's expected that this infrastructure-only change has no real
impact on UIs, and this is assumed to be the case as per the passing
tests.